# openapi.yaml
openapi: 3.0.3
info:
  title: API Sổ thu chi cá nhân
  version: 2.0.0
  description: REST API cho TASK_00030. Mọi response JSON sử dụng envelope success/message/data hoặc success/error.
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Danh sách giao dịch có lọc và phân trang
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { in: query, name: walletId, schema: { type: string, format: uuid } }
        - { in: query, name: categoryId, schema: { type: string, format: uuid } }
        - { in: query, name: type, schema: { type: string, enum: [INCOME, EXPENSE, TRANSFER] } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: keyword, schema: { type: string } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
    post:
      tags: [Transactions]
      summary: Ghi giao dịch
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TransactionInput' } } }
      responses: { '201': { $ref: '#/components/responses/Success' } }
...
  /transactions/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Transactions]
      summary: Chi tiết giao dịch
      responses: { '200': { $ref: '#/components/responses/Success' } }
    patch:
      tags: [Transactions]
      summary: Sửa giao dịch
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TransactionInput' } } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
    delete:
      tags: [Transactions]
      summary: Xóa giao dịch
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /transactions/{id}/receipts:
    post:
      tags: [Transactions]
      summary: Upload một hóa đơn JPG, PNG hoặc PDF
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties: { file: { type: string, format: binary } }
      responses: { '201': { $ref: '#/components/responses/Success' } }
  /transactions/{transactionId}/receipts/{receiptId}:
    get:
      tags: [Transactions]
      summary: Tải hóa đơn (có kiểm tra quyền sở hữu)
      parameters:
        - { in: path, name: transactionId, required: true, schema: { type: string, format: uuid } }
        - { in: path, name: receiptId, required: true, schema: { type: string, format: uuid } }
      responses: { '200': { description: Nội dung tệp } }
    delete:
      tags: [Transactions]
      summary: Xóa riêng một hóa đơn
      parameters:
        - { in: path, name: transactionId, required: true, schema: { type: string, format: uuid } }
        - { in: path, name: receiptId, required: true, schema: { type: string, format: uuid } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
  /transactions/bulk:
    post:
      tags: [Transactions]
      summary: Xóa, khôi phục hoặc đối soát nhiều giao dịch
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [ids, action], properties: { ids: { type: array, items: { type: string, format: uuid } }, action: { type: string, enum: [DELETE, RESTORE, RECONCILE] } } } } } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
  /transactions/{id}/restore:
    post:
      tags: [Transactions]
      summary: Khôi phục giao dịch
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /transactions/{id}/splits:
    put:
      tags: [Transactions]
      summary: Chia giao dịch thành nhiều phần
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Success:
      description: Thành công
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: Thành công. }
        data: { nullable: true }
        meta: { type: object, nullable: true }
...
    TransactionInput:
      type: object
      required: [walletId, type, amount, occurredAt]
      properties:
        walletId: { type: string, format: uuid }
        destinationWalletId: { type: string, format: uuid, nullable: true }
        categoryId: { type: string, format: uuid, nullable: true }
        type: { type: string, enum: [INCOME, EXPENSE, TRANSFER] }
        amount: { type: number, minimum: 0, example: 50000 }
        occurredAt: { type: string, format: date-time }
        note: { type: string, nullable: true }
...
security:
  - bearerAuth: []